Fix for macros without parameters
Previously macros such as `@__FILE__` would not have permalinks since
they did not include parenthesis. Not only was this annoying for cross-
referencing but it made searching difficult as searching for `@__FILE__`
would result in no search results.

Note: Parenthesis will automatically be added to macros without them in
the rendered documentation.